Move defineProgramName function out of time file
