+++ services/ui-src/src/utils/other/time.ts
@@ -1,5 +1,5 @@
 import { utcToZonedTime, zonedTimeToUtc } from "date-fns-tz";
-import { Choice, DateShape, DropdownChoice, TimeShape } from "types";
+import { DateShape, TimeShape } from "types";
 
 export const midnight: TimeShape = { hour: 0, minute: 0, second: 0 };
 export const oneSecondToMidnight: TimeShape = {
@@ -161,30 +161,4 @@ export const calculateDueDate = (date: string) => {
 export const calculateRemainingSeconds = (expiresAt?: any) => {
   if (!expiresAt) return 0;
   return (new Date(expiresAt).valueOf() - Date.now()) / 1000;
-};
-
-export const defineProgramName = (
-  newOrExistingProgram: Choice[],
-  existingProgramNameSelection?: DropdownChoice,
-  newProgramName?: string
-) => {
-  //TODO - Seek clarity on this function and how it should operate.
-  if (!newOrExistingProgram?.[0] || !newOrExistingProgram?.[0].value)
-    throw new Error(
-      "Program name radio field was not selected as an existing or new report"
-    );
-
-  if (newOrExistingProgram[0].value === "Existing program") {
-    if (!existingProgramNameSelection?.value)
-      throw new Error(
-        "Program name dropdown selection did not have a value. Please double check the field is working properly."
-      );
-    return existingProgramNameSelection.value;
-  } else if (newOrExistingProgram[0].value === "Add new program") {
-    return newProgramName;
-  } else {
-    throw new Error(
-      "A choice was made in the program name selection field that isn't supported. Please add your choice to this function (defineProgramName) or fix the typo in the addEditModalJson file."
-    );
-  }
 };
